Pin swagger-parser version
Pins the swagger-parser version, this may cause a delay between when a
fixed version comes out and a new version of this is released for new
users, but it means that we can push out an update of this plugin and get
_current_ users updated to the latest version, as Atom has no simple way
to update something like this for an installed package.

GreenKeeper will send us a PR for every new version of `swagger-parser`
with this change.