Fix a race condition
If the editor text changed since the lint was initially triggered we could
end up attempting to mark locations that are no longer valid. This puts in
a check that if this has hapened we simply tell Linter not to update the
current set of messages that it has. As this editor has lint-on-fly
enabled, whatever changes that caused this to happen will have triggered
their own `lint()` call, so nothing is lost, just slightly delayed.