Handle undefined ranges
If the findRangeRecursive function was unable to determine a range for
an error message, use the entire first line of the file as the range
reported. This isn't exactly ideal, but doing this in the same way
errors like this are handled in other providers would require a large
rewrite.